@@ -2,6 +2,7 @@ using System.Diagnostics;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using vassago.Models;
+using vassago.ProtocolInterfaces.DiscordInterface;
 
 namespace vassago.Controllers.api;
 
@@ -10,49 +11,50 @@ namespace vassago.Controllers.api;
 public class ChannelsController : ControllerBase
 {
     private readonly ILogger<ChannelsController> _logger;
-    private readonly ChattingContext _db;
 
-    public ChannelsController(ILogger<ChannelsController> logger, ChattingContext db)
+    public ChannelsController(ILogger<ChannelsController> logger)
     {
         _logger = logger;
-        _db = db;
     }
 
     [HttpGet("{id}")]
     [Produces("application/json")]
     public Channel Get(Guid id)
     {
-        return _db.Find<Channel>(id);
+        return Rememberer.ChannelDetail(id);
     }
 
     [HttpPatch]
     [Produces("application/json")]
     public IActionResult Patch([FromBody] Channel channel)
     {
-        var fromDb = _db.Channels.Find(channel.Id);
+        var fromDb = Rememberer.ChannelDetail(channel.Id);
         if (fromDb == null)
 		{
 			_logger.LogError($"attempt to update channel {channel.Id}, not found");
 			return NotFound();
         }
+        else
+        {
+            _logger.LogDebug($"patching {channel.DisplayName} (id: {channel.Id})");
+        }
 		//settable values: lewdness filter level, meanness filter level. maybe i could decorate them... 
 		fromDb.LewdnessFilterLevel = channel.LewdnessFilterLevel;
 		fromDb.MeannessFilterLevel = channel.MeannessFilterLevel;
-		_db.SaveChanges();
+		Rememberer.RememberChannel(fromDb);
         return Ok(fromDb);
     }
     [HttpDelete]
     [Produces("application/json")]
     public IActionResult Delete([FromBody] Channel channel)
     {
-        var fromDb = _db.Channels.Find(channel.Id);
+        var fromDb = Rememberer.ChannelDetail(channel.Id);
         if (fromDb == null)
 		{
 			_logger.LogError($"attempt to delete channel {channel.Id}, not found");
 			return NotFound();
         }
         deleteChannel(fromDb);
-        _db.SaveChanges();
         return Ok();
     }
     private void deleteChannel(Channel channel)
@@ -73,21 +75,16 @@ public class ChannelsController : ControllerBase
             }
         }
 
-        if(channel.Messages?.Count > 0)
-        {
-            _db.Remove(channel.Messages);
-        }
-
-        _db.Remove(channel);
+        Rememberer.ForgetChannel(channel);
     }
     private void deleteAccount(Account account)
     {
         var user = account.IsUser;
         var usersOnlyAccount = user.Accounts?.Count == 1;
         
-        _db.Remove(account);
+        Rememberer.ForgetAccount(account);
         
         if(usersOnlyAccount)
-            _db.Users.Remove(user);
+            Rememberer.ForgetUser(user);
     }
 }
