@@ -1,8 +1,7 @@
 ﻿@{
     ViewData["Title"] = "Home Page";
 }
-<div id="tree"></div>
-tree above.
+<div id="tree">tree here</div>
 
 @section Scripts{
     <script type="text/javascript">
