@@ -1,4 +1,4 @@
-﻿@model ErrorPageViewModel
+﻿@model vassago.WebInterface.Models.ErrorPageViewModel
 @{
     ViewData["Title"] = "Error";
 }
